Move LevelUpdate serialize deseriaize tests to validator
